requirements.txt
django~=5.1.1
environs~= 11.0.0
pillow~=10.4.0
djangorestframework~=3.15.2
django-jazzmin~=3.0.0